requirements.txt
-e .
absl-py==0.9.0
astor==0.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
event-service-utils==1.2.6
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.0; python_version < '3.8'
jaeger-client==4.1.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
logzero==1.5.0
markdown==3.2.2; python_version >= '3.5'
numpy==1.16.4
opencv-python-headless==4.0.0.21
opentracing==2.1.0
pillow==5.4.1
prometheus-client==0.7.1
protobuf==3.12.2
python-decouple==3.1
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadloop==1.0.2
thrift==0.13.0
tornado==5.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
walrus==0.7.1
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.34.2; python_version >= '3'
wrapt==1.12.1
zipp==3.1.0; python_version >= '3.6'
